Tools: enforce SPDX check
